+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/component/Dialog.kt
@@ -1,12 +1,13 @@
 package co.electriccoin.zcash.ui.design.component
 
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -61,7 +62,7 @@ fun AppAlertDialog(
     properties: DialogProperties = DialogProperties()
 ) {
     AlertDialog(
-        shape = RectangleShape,
+        shape = RoundedCornerShape(corner = CornerSize(ZcashTheme.dimens.regularRippleEffectCorner)),
         onDismissRequest = onDismissRequest?.let { onDismissRequest } ?: {},
         confirmButton = {
             confirmButtonText?.let {