+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/home/AndroidHome.kt
@@ -18,7 +18,6 @@ import cash.z.ecc.android.sdk.model.ZecSend
 import co.electriccoin.zcash.ui.MainActivity
 import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.common.compose.RestoreScreenBrightness
-import co.electriccoin.zcash.ui.common.model.VersionInfo
 import co.electriccoin.zcash.ui.common.model.WalletRestoringState
 import co.electriccoin.zcash.ui.common.model.WalletSnapshot
 import co.electriccoin.zcash.ui.common.viewmodel.HomeViewModel
@@ -56,13 +55,6 @@ internal fun MainActivity.WrapHome(
 
     val isRestoringInitialWarningSeen = homeViewModel.isRestoringInitialWarningSeen.collectAsStateWithLifecycle().value
 
-    // Detailed sync status info is used if set in configuration or if the app is built as debuggable
-    // (i.e. mainly in development)
-    val isDetailedSyncStatus =
-        homeViewModel.isDetailedSyncStatus.collectAsStateWithLifecycle().value.run {
-            this ?: false || VersionInfo.new(this@WrapHome).isDebuggable
-        }
-
     val walletSnapshot = walletViewModel.walletSnapshot.collectAsStateWithLifecycle().value
 
     val walletRestoringState = walletViewModel.walletRestoringState.collectAsStateWithLifecycle().value
@@ -98,7 +90,6 @@ internal fun MainActivity.WrapHome(
         goSettings = goSettings,
         goMultiTrxSubmissionFailure = goMultiTrxSubmissionFailure,
         homeScreenIndex = homeScreenIndex,
-        isDetailedSyncStatus = isDetailedSyncStatus,
         isKeepScreenOnWhileSyncing = isKeepScreenOnWhileSyncing,
         isShowingRestoreInitDialog = isShowingRestoreInitDialog,
         onPageChange = {
@@ -120,7 +111,6 @@ internal fun WrapHome(
     goScan: () -> Unit,
     goSendConfirmation: (ZecSend) -> Unit,
     homeScreenIndex: HomeScreenIndex,
-    isDetailedSyncStatus: Boolean,
     isKeepScreenOnWhileSyncing: Boolean?,
     isShowingRestoreInitDialog: Boolean,
     onPageChange: (HomeScreenIndex) -> Unit,
@@ -203,7 +193,6 @@ internal fun WrapHome(
                 screenContent = {
                     WrapBalances(
                         activity = activity,
-                        isDetailedSyncStatus = isDetailedSyncStatus,
                         goSettings = goSettings,
                         goMultiTrxSubmissionFailure = goMultiTrxSubmissionFailure
                     )